#2: Move types to formatter package
