+++ formatter/config.go
@@ -2,14 +2,12 @@ package formatter
 
 import (
 	"io"
-
-	"github.com/vdjagilev/nmap-formatter/types"
 )
 
 type Config struct {
 	Writer        io.Writer
-	OutputFormat  types.OutputFormat
-	InputFile     types.InputFile
-	OutputFile    types.OutputFile
-	OutputOptions types.OutputOptions
+	OutputFormat  OutputFormat
+	InputFile     InputFile
+	OutputFile    OutputFile
+	OutputOptions OutputOptions
 }
