+++ formatter/formatter_csv.go
@@ -2,8 +2,6 @@ package formatter
 
 import (
 	"encoding/csv"
-
-	"github.com/vdjagilev/nmap-formatter/types"
 )
 
 type CSVFormatter struct {
@@ -11,12 +9,12 @@ type CSVFormatter struct {
 }
 
 // Format the data to CSV and output it to appropriate io.Writer
-func (f *CSVFormatter) Format(td *types.TemplateData) (err error) {
+func (f *CSVFormatter) Format(td *TemplateData) (err error) {
 	return csv.NewWriter(f.Config.Writer).WriteAll(f.convert(td))
 }
 
 // convert uses NMAPRun struct to convert all data to [][]string type
-func (f *CSVFormatter) convert(td *types.TemplateData) (data [][]string) {
+func (f *CSVFormatter) convert(td *TemplateData) (data [][]string) {
 	data = append(data, []string{"IP", "Port", "Protocol", "State", "Service", "Reason", "Product", "Version", "Extra info"})
 	for _, host := range td.NMAPRun.Host {
 		for j, port := range host.Ports.Port {
