+++ formatter/formatter_json.go
@@ -3,8 +3,6 @@ package formatter
 import (
 	"bytes"
 	"encoding/json"
-
-	"github.com/vdjagilev/nmap-formatter/types"
 )
 
 type JSONFormatter struct {
@@ -12,7 +10,7 @@ type JSONFormatter struct {
 }
 
 // Format the data and output it to appropriate io.Writer
-func (f *JSONFormatter) Format(td *types.TemplateData) (err error) {
+func (f *JSONFormatter) Format(td *TemplateData) (err error) {
 	jsonData := new(bytes.Buffer)
 	err = json.NewEncoder(jsonData).Encode(td.NMAPRun)
 	if err != nil {
